[pt2] support grad_and_value
