+++ torch/_functorch/deprecated.py
@@ -68,7 +68,7 @@ def grad(func: Callable, argnums: argnums_t = 0, has_aux: bool = False) -> Calla
 
 def grad_and_value(func: Callable, argnums: argnums_t = 0, has_aux: bool = False) -> Callable:
     warn_deprecated('grad_and_value')
-    return _impl.grad_and_value(func, argnums, has_aux)
+    return apis.grad_and_value(func, argnums, has_aux)
 
 def vjp(func: Callable, *primals, has_aux: bool = False):
     warn_deprecated('vjp')
@@ -111,7 +111,7 @@ def combine_state_for_ensemble(models):
 
 setup_docs(vmap, _vmap_impl.vmap, 'torch.vmap')
 setup_docs(grad, apis.grad)
-setup_docs(grad_and_value)
+setup_docs(grad_and_value, apis.grad_and_value)
 setup_docs(vjp)
 setup_docs(jvp)
 setup_docs(jacrev)
