+++ torch/func/__init__.py
@@ -1,5 +1,4 @@
 from torch._functorch.eager_transforms import (
-    grad_and_value,
     vjp,
     jvp,
     jacrev,
@@ -8,7 +7,7 @@ from torch._functorch.eager_transforms import (
     functionalize,
     linearize
 )
-from torch._functorch.apis import grad
+from torch._functorch.apis import grad, grad_and_value
 from torch._functorch.functional_call import functional_call, stack_module_state
 from torch._functorch.batch_norm_replacement import replace_all_batch_norm_modules_
 from torch._functorch.vmap import vmap
